+++ scenes/settings.c
@@ -24,6 +24,13 @@ static void emerg_changed(VariableItem* item) {
     app->emergency_vehicle_detect = (idx == 1);
 }
 
+static void nag_killer_changed(VariableItem* item) {
+    TeslaFSDApp* app = variable_item_get_context(item);
+    uint8_t idx = variable_item_get_current_value_index(item);
+    variable_item_set_current_value_text(item, toggle_text[idx]);
+    app->nag_killer = (idx == 1);
+}
+
 void tesla_fsd_scene_settings_on_enter(void* context) {
     TeslaFSDApp* app = context;
     VariableItemList* list = app->var_item_list;
@@ -42,6 +49,10 @@ void tesla_fsd_scene_settings_on_enter(void* context) {
     item = variable_item_list_add(list, "Emerg. Vehicle", 2, emerg_changed, app);
     variable_item_set_current_value_index(item, app->emergency_vehicle_detect ? 1 : 0);
     variable_item_set_current_value_text(item, toggle_text[app->emergency_vehicle_detect ? 1 : 0]);
+
+    item = variable_item_list_add(list, "Nag Killer", 2, nag_killer_changed, app);
+    variable_item_set_current_value_index(item, app->nag_killer ? 1 : 0);
+    variable_item_set_current_value_text(item, toggle_text[app->nag_killer ? 1 : 0]);
 
     view_dispatcher_switch_to_view(app->view_dispatcher, TeslaFSDViewVarItemList);
 }